Move useInterval requested by @bryancodazen
