+++ beta-src/src/hooks/useInterval.ts
